+++ src/components/noticeRegister/Label.tsx
