+++ src/pages/notice/register/index.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
 import { useRouter } from "next/router";
 
-import { Input } from "@/components/NoticeRegister/noticeRegisterInput";
-import { Label } from "@/components/NoticeRegister/noticeRegisterLabel";
+import { Input } from "@/components/noticeRegister/Input";
+import { Label } from "@/components/noticeRegister/Label";
 
 function NoticeRegister() {
   const router = useRouter();
@@ -29,39 +29,41 @@ function NoticeRegister() {
                 />
               </div>
               <div className="flex h-[52.3rem] w-full flex-col justify-end gap-[2rem]">
-                <form className="flex h-[9.2rem] w-full flex-col items-start gap-[0.8rem]">
-                  <Label htmlFor="시급">시급*</Label>
-                  <div className="relative inline-block w-full">
-                    <Input type="text" id="hourly wage" />
-                    <span className="absolute inset-y-0 right-0 flex items-center pr-[2rem] text-[1.6rem] font-normal not-italic leading-[2.6rem] text-black">
-                      원
-                    </span>
+                <form>
+                  <div className="flex h-[9.2rem] w-full flex-col items-start gap-[0.8rem]">
+                    <Label htmlFor="시급">시급*</Label>
+                    <div className="relative inline-block w-full">
+                      <Input type="text" id="hourly wage" />
+                      <span className="absolute inset-y-0 right-0 flex items-center pr-[2rem] text-[1.6rem] font-normal not-italic leading-[2.6rem] text-black">
+                        원
+                      </span>
+                    </div>
                   </div>
-                </form>
-                <form className="flex h-[9.2rem] w-full flex-col items-start gap-[0.8rem]">
-                  <Label htmlFor="시작 일시">시작 일시*</Label>
-                  <Input type="text" id="begin" />
-                </form>
-                <form className="flex h-[9.2rem] w-full flex-col items-start gap-[0.8rem]">
-                  <Label htmlFor="업무 시간">업무 시간*</Label>
-                  <div className="relative inline-block w-full">
+                  <div className="flex h-[9.2rem] w-full flex-col items-start gap-[0.8rem]">
+                    <Label htmlFor="시작 일시">시작 일시*</Label>
+                    <Input type="text" id="begin" />
+                  </div>
+                  <div className="flex h-[9.2rem] w-full flex-col items-start gap-[0.8rem]">
+                    <Label htmlFor="업무 시간">업무 시간*</Label>
+                    <div className="relative inline-block w-full">
+                      <Input
+                        type="text"
+                        className="placeholder:italic"
+                        id="business hours"
+                      />
+                      <span className="absolute inset-y-0 right-0 flex items-center pr-[2rem] text-[1.6rem] font-normal not-italic leading-[2.6rem] text-black">
+                        시간
+                      </span>
+                    </div>
+                  </div>
+                  <div className="flex h-[18.7rem] w-full flex-col items-start gap-[0.8rem]">
+                    <Label htmlFor="업무 시간">공고 설명</Label>
                     <Input
                       type="text"
-                      className="placeholder:italic"
-                      id="business hours"
+                      id="explanation"
+                      className="flex h-[15.3rem] w-full items-center self-stretch overflow-auto rounded-md border border-input px-[2rem] py-[1.6rem] text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
                     />
-                    <span className="absolute inset-y-0 right-0 flex items-center pr-[2rem] text-[1.6rem] font-normal not-italic leading-[2.6rem] text-black">
-                      시간
-                    </span>
                   </div>
-                </form>
-                <form className="flex h-[18.7rem] w-full flex-col items-start gap-[0.8rem]">
-                  <Label htmlFor="업무 시간">공고 설명</Label>
-                  <Input
-                    type="text"
-                    id="explanation"
-                    className="flex h-[15.3rem] w-full items-center self-stretch overflow-auto rounded-md border border-input px-[2rem] py-[1.6rem] text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                  />
                 </form>
               </div>
               <div>
